test: organize test
